'自动同步'

+++ model/src/main/java/com/redfish/common/query/model/service/CountQueryService.java
@@ -11,6 +11,6 @@ public interface CountQueryService {
      * @param countQueryParam
      * @return
      */
-    Long count(CountQueryParam countQueryParam);
+    Integer count(CountQueryParam countQueryParam);
 
 }
